+++ docs/changelog.rst
@@ -4,7 +4,10 @@ Changelog
 
 * :release:`1.0.0 <future>`
 
-* :feature:`109` Add the CAEKL mutual information
+* :feature:`11` Add the exact common information.
+* :feature:`9` Add the Wyner common information.
+* :feature:`111` Add the functional common information.
+* :feature:`109` Add the CAEKL mutual information.
 * :feature:`108` Add the multivariate entropy triangle.
 * :feature:`63` Add scalar operations for ScalarDistributions.
 * :feature:`96` Add the marginal utility of information.